Conflicts:
	DataFormats/L1TrackTrigger/interface/TTTrack_TrackWord.h
	L1Trigger/L1TTrackMatch/test/L1TrackObjectNtupleMaker.cc
	L1Trigger/L1TTrackMatch/test/L1TrackObjectNtupleMaker_cfg.py
	L1Trigger/L1TTrackMatch/test/vtxPlotter.C
	L1Trigger/TrackFindingTracklet/plugins/L1FPGATrackProducer.cc
	L1Trigger/VertexFinder/data/Quantised_model_prune_iteration_9_associationModelgraph.pb
	L1Trigger/VertexFinder/data/Quantised_model_prune_iteration_9_patternModelgraph.pb
	L1Trigger/VertexFinder/data/Quantised_model_prune_iteration_9_weightModelgraph.pb
	L1Trigger/VertexFinder/src/VertexFinder.cc

+++ DataFormats/L1TrackTrigger/interface/TTTrack_TrackWord.h
@@ -225,7 +225,7 @@ public:
   double getChi2RZ() const { return chi2RZBins[getChi2RZBits()]; }
   double getBendChi2() const { return bendChi2Bins[getBendChi2Bits()]; }
   unsigned int getHitPattern() const { return getHitPatternBits(); }
-  unsigned int getMVAQuality() const { return mvaQualityBins[getMVAQualityBits()]; }
+  double getMVAQuality() const { return mvaQualityBins[getMVAQualityBits()]; }
   unsigned int getMVAOther() const { return getMVAOtherBits(); }
 
   // ----------member functions (setters) ------------
